refactor(config): necessary code clean up
