+++ tsconfig.json
@@ -1,4 +1,6 @@
 {
-  // https://v3.nuxtjs.org/concepts/typescript
   "extends": "./.nuxt/tsconfig.json",
+  "compilerOptions": {
+    "strictNullChecks": true
+  }
 }
